requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
autopep8==2.3.2
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
fastapi==0.115.11
FastAPI-SQLAlchemy==0.2.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.5
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
typing_extensions==4.12.2
uvicorn==0.34.0
